Revert "refactor: optimize imports and refactor ProfileForm components (#6223)"
This reverts commit 1254eede8d03c9140e92368f0f12be8117d35320.
