--- docs/docs/auto-docs/constants/userRoles/type-aliases/UserRole.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Type Alias: UserRole
-
-> **UserRole** = *typeof* [`USER_ROLES`](../variables/USER_ROLES.md)\[keyof *typeof* [`USER_ROLES`](../variables/USER_ROLES.md)\]
-
-Defined in: [src/constants/userRoles.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/constants/userRoles.ts#L10)
