--- docs/docs/auto-docs/constants/userRoles/variables/USER_ROLES.md
@@ -1,21 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: USER\_ROLES
-
-> `const` **USER\_ROLES**: `object`
-
-Defined in: [src/constants/userRoles.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/constants/userRoles.ts#L4)
-
-Constants for User Roles.
-
-## Type Declaration
-
-### ADMIN
-
-> `readonly` **ADMIN**: `"administrator"` = `'administrator'`
-
-### USER
-
-> `readonly` **USER**: `"regular"` = `'regular'`
