+++ docs/docs/auto-docs/screens/LoginPage/LoginPage/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(): `Element`
 
-Defined in: [src/screens/LoginPage/LoginPage.tsx:82](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/LoginPage/LoginPage.tsx#L82)
+Defined in: [src/screens/LoginPage/LoginPage.tsx:51](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/LoginPage/LoginPage.tsx#L51)
 
 ## Returns
