--- docs/docs/auto-docs/shared-components/ProfileForm/ProfileContactInfo/variables/default.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: default
-
-> `const` **default**: `React.FC`\<[`IProfileContactInfoProps`](../../types/interfaces/IProfileContactInfoProps.md)\>
-
-Defined in: [src/shared-components/ProfileForm/ProfileContactInfo.tsx:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/ProfileContactInfo.tsx#L15)
