+++ docs/docs/auto-docs/shared-components/ProfileForm/ProfileForm/functions/getLanguageName.md
@@ -6,7 +6,7 @@
 
 > **getLanguageName**(`code`): `string`
 
-Defined in: [src/shared-components/ProfileForm/ProfileForm.tsx:415](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/ProfileForm.tsx#L415)
+Defined in: [src/shared-components/ProfileForm/ProfileForm.tsx:816](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/ProfileForm.tsx#L816)
 
 ## Parameters
