+++ docs/docs/auto-docs/shared-components/ProfileForm/ProfileForm/variables/default.md
@@ -6,4 +6,4 @@
 
 > `const` **default**: `React.FC`
 
-Defined in: [src/shared-components/ProfileForm/ProfileForm.tsx:106](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/ProfileForm.tsx#L106)
+Defined in: [src/shared-components/ProfileForm/ProfileForm.tsx:115](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/ProfileForm.tsx#L115)
