--- docs/docs/auto-docs/shared-components/ProfileForm/ProfilePersonalDetails/variables/default.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: default
-
-> `const` **default**: `React.FC`\<[`IProfilePersonalDetailsProps`](../../types/interfaces/IProfilePersonalDetailsProps.md)\>
-
-Defined in: [src/shared-components/ProfileForm/ProfilePersonalDetails.tsx:28](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/ProfilePersonalDetails.tsx#L28)
