--- docs/docs/auto-docs/shared-components/ProfileForm/types/interfaces/IProfileContactInfoProps.md
@@ -1,45 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: IProfileContactInfoProps
-
-Defined in: [src/shared-components/ProfileForm/types.ts:29](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L29)
-
-## Properties
-
-### formState
-
-> **formState**: [`IProfileFormState`](IProfileFormState.md)
-
-Defined in: [src/shared-components/ProfileForm/types.ts:30](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L30)
-
-***
-
-### handleFieldChange()
-
-> **handleFieldChange**: (`fieldName`, `value`) => `void`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:31](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L31)
-
-#### Parameters
-
-##### fieldName
-
-keyof [`IProfileFormState`](IProfileFormState.md)
-
-##### value
-
-`string`
-
-#### Returns
-
-`void`
-
-***
-
-### userEmail?
-
-> `optional` **userEmail**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:35](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L35)
