--- docs/docs/auto-docs/shared-components/ProfileForm/types/interfaces/IProfileFormState.md
@@ -1,178 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: IProfileFormState
-
-Defined in: [src/shared-components/ProfileForm/types.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L5)
-
-Interface definition for the Profile Form State.
-This ensures consistency across ProfileForm and its subcomponents.
-
-## Properties
-
-### addressLine1
-
-> **addressLine1**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L6)
-
-***
-
-### addressLine2
-
-> **addressLine2**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L7)
-
-***
-
-### avatar
-
-> **avatar**: `File`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L11)
-
-***
-
-### avatarURL
-
-> **avatarURL**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L12)
-
-***
-
-### birthDate
-
-> **birthDate**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L8)
-
-***
-
-### city
-
-> **city**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L10)
-
-***
-
-### countryCode
-
-> **countryCode**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L13)
-
-***
-
-### description
-
-> **description**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:14](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L14)
-
-***
-
-### educationGrade
-
-> **educationGrade**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L15)
-
-***
-
-### emailAddress
-
-> **emailAddress**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L9)
-
-***
-
-### employmentStatus
-
-> **employmentStatus**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L16)
-
-***
-
-### homePhoneNumber
-
-> **homePhoneNumber**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L17)
-
-***
-
-### maritalStatus
-
-> **maritalStatus**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:18](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L18)
-
-***
-
-### mobilePhoneNumber
-
-> **mobilePhoneNumber**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L19)
-
-***
-
-### name
-
-> **name**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L20)
-
-***
-
-### natalSex
-
-> **natalSex**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:21](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L21)
-
-***
-
-### naturalLanguageCode
-
-> **naturalLanguageCode**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L22)
-
-***
-
-### password
-
-> **password**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:23](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L23)
-
-***
-
-### postalCode
-
-> **postalCode**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:24](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L24)
-
-***
-
-### state
-
-> **state**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:25](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L25)
-
-***
-
-### workPhoneNumber
-
-> **workPhoneNumber**: `string`
-
-Defined in: [src/shared-components/ProfileForm/types.ts:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/ProfileForm/types.ts#L26)
