+++ public/locales/en/translation.json
@@ -1247,8 +1247,7 @@
     "unassignUserTagMessage": "Do you want to remove the tag from this user?",
     "successfullyUnassigned": "Tag unassigned from user",
     "tagsAssigned": "Tags Assigned",
-    "noTagsAssigned": "No Tags Assigned",
-    "passwordPlaceholder": "Leave blank to keep current password"
+    "noTagsAssigned": "No Tags Assigned"
   },
   "people": {
     "title": "People",
