+++ public/locales/hi/translation.json
@@ -1193,8 +1193,7 @@
     "unassignUserTagMessage": "क्या आप इस उपयोगकर्ता से टैग हटाना चाहते हैं?",
     "successfullyUnassigned": "उपयोगकर्ता से टैग हटा दिया गया",
     "tagsAssigned": "टैग सौंपे गए",
-    "noTagsAssigned": "कोई टैग सौंपा नहीं गया",
-    "passwordPlaceholder": "वर्तमान पासवर्ड रखने के लिए खाली छोड़ें"
+    "noTagsAssigned": "कोई टैग सौंपा नहीं गया"
   },
   "people": {
     "title": "लोग",
