+++ public/locales/zh/translation.json
@@ -1193,8 +1193,7 @@
     "unassignUserTagMessage": "您想从此用户中删除标签吗？",
     "successfullyUnassigned": "标签已从用户中取消分配",
     "tagsAssigned": "已分配标签",
-    "noTagsAssigned": "未分配标签",
-    "passwordPlaceholder": "留空以保留当前密码"
+    "noTagsAssigned": "未分配标签"
   },
   "userLogin": {
     "login": "登录",
