+++ src/components/DynamicDropDown/DynamicDropDown.spec.tsx
@@ -24,10 +24,7 @@ const renderComponent = (props: Record<string, unknown> = {}) => {
   return render(
     <BrowserRouter>
       <I18nextProvider i18n={i18nForTest}>
-        <DynamicDropDown
-          {...defaultProps}
-          formState={defaultProps.formState as Record<string, unknown>}
-        />
+        <DynamicDropDown {...defaultProps} />
       </I18nextProvider>
     </BrowserRouter>,
   );
