+++ src/screens/LoginPage/LoginPage.tsx
@@ -1,40 +1,11 @@
 /**
- * LoginPage.tsx
+ * LoginPage component.
  *
- * This file contains the implementation of the Login and Registration page for the Talawa Admin application.
- * It includes functionality for user authentication, password validation, reCAPTCHA verification, and organization selection.
- * The page supports both admin and user roles and provides localization support.
+ * Provides login and registration flows with validation, reCAPTCHA, and
+ * organization selection. Supports admin and user roles with localization.
  *
- * Dependencies:
- * - react
- * - react-router-dom
- * - react-bootstrap
- * - react-google-recaptcha
- * - `@apollo/client`
- * - `@mui/icons-material`
- * - `@mui/material`
- * - NotificationToast
- * - i18next
- * - utils/errorHandler
- * - utils/useLocalstorage
- * - utils/useSession
- * - utils/i18n
- * - GraphQl/Mutations/mutations
- * - GraphQl/Queries/Queries
- * - components/ChangeLanguageDropdown/ChangeLanguageDropDown
- * - components/LoginPortalToggle/LoginPortalToggle
- * - assets/svgs/palisadoes.svg
- * - assets/svgs/talawa.svg
- *
- * The `loginPage` component renders a login and registration interface with the following features:
- * - Login and registration forms with validation.
- * - Password strength checks and visibility toggles.
- * - reCAPTCHA integration for bot prevention.
- * - Organization selection using an autocomplete dropdown.
- * - Social media links and community branding.
- * - Role-based navigation for admin and user.
- *
- * @returns The rendered login and registration page.
+ * @remarks
+ * Includes password strength checks, social links, and community branding.
  *
  * @example
  * ```tsx
@@ -42,8 +13,7 @@
  * ```
  */
 import { useQuery, useMutation, useLazyQuery } from '@apollo/client';
-import Check from '@mui/icons-material/Check';
-import Clear from '@mui/icons-material/Clear';
+import { Check, Clear } from '@mui/icons-material';
 import type { ChangeEvent } from 'react';
 import React, { useEffect, useRef, useState } from 'react';
 import { Form } from 'react-bootstrap';
@@ -74,8 +44,7 @@ import useLocalStorage from 'utils/useLocalstorage';
 import { socialMediaLinks } from '../../constants';
 import styles from './LoginPage.module.css';
 import type { InterfaceQueryOrganizationListObject } from 'utils/interfaces';
-import Autocomplete from '@mui/material/Autocomplete';
-import TextField from '@mui/material/TextField';
+import { Autocomplete, TextField } from '@mui/material';
 import useSession from 'utils/useSession';
 import i18n from 'utils/i18n';
