--- src/shared-components/ProfileForm/ProfileContactInfo.spec.tsx
@@ -1,177 +1,0 @@
-import { vi } from 'vitest';
-import React from 'react';
-import { render, screen, fireEvent, within } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import ProfileContactInfo from './ProfileContactInfo';
-import { IProfileFormState } from './types';
-import { countryOptions } from 'utils/formEnumFields';
-
-// Mock translation
-vi.mock('react-i18next', () => ({
-  useTranslation: (_ns: string) => {
-    return {
-      t: (key: string) => key,
-      i18n: {
-        changeLanguage: vi.fn(),
-      },
-    };
-  },
-}));
-
-describe('ProfileContactInfo', () => {
-  afterEach(() => {
-    vi.clearAllMocks();
-  });
-
-  const mockHandleFieldChange = vi.fn();
-  const mockFormState: IProfileFormState = {
-    mobilePhoneNumber: '1234567890',
-    workPhoneNumber: '0987654321',
-    homePhoneNumber: '1122334455',
-    addressLine1: '123 Main St',
-    addressLine2: 'Apt 4B',
-    postalCode: '90210',
-    city: 'Beverly Hills',
-    state: 'CA',
-    countryCode: 'US',
-    // Add other required fields with dummy values
-    name: 'Test User',
-    natalSex: 'Male',
-    birthDate: '1990-01-01',
-    educationGrade: 'Bachelor',
-    employmentStatus: 'Employed',
-    maritalStatus: 'Single',
-    password: '',
-    description: 'A test user',
-    avatarURL: '',
-    emailAddress: 'test@example.com',
-    avatar: null,
-    naturalLanguageCode: 'en',
-  };
-
-  const defaultProps = {
-    formState: mockFormState,
-    handleFieldChange: mockHandleFieldChange,
-    userEmail: 'test@example.com',
-  };
-
-  it('renders all contact info fields with correct translation keys and placeholders', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-
-    // Headings and Labels
-    expect(screen.getByText('contactInfoHeading')).toBeInTheDocument();
-    expect(screen.getByLabelText('email')).toBeInTheDocument();
-    expect(screen.getByLabelText('mobilePhoneNumber')).toBeInTheDocument();
-    expect(screen.getByLabelText('workPhoneNumber')).toBeInTheDocument();
-    expect(screen.getByLabelText('homePhoneNumber')).toBeInTheDocument();
-    expect(screen.getByLabelText('addressLine1')).toBeInTheDocument();
-    expect(screen.getByLabelText('addressLine2')).toBeInTheDocument();
-    expect(screen.getByLabelText('postalCode')).toBeInTheDocument();
-    expect(screen.getByLabelText('city')).toBeInTheDocument();
-    expect(screen.getByLabelText('state')).toBeInTheDocument();
-    expect(screen.getByLabelText('country')).toBeInTheDocument();
-
-    // Check placeholders (using the translation key passed to tCommon)
-    // Note: In the component tCommon is used for placeholders like "example" or "enterCityName"
-    expect(screen.getByPlaceholderText('email')).toBeInTheDocument(); // placeholder={tCommon('email')}
-    expect(screen.getByPlaceholderText('enterCityName')).toBeInTheDocument();
-    expect(screen.getByPlaceholderText('enterStateName')).toBeInTheDocument();
-  });
-
-  it('renders user email as disabled', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-    const emailInput = screen.getByLabelText('email');
-    expect(emailInput).toBeDisabled();
-    expect(emailInput).toHaveValue('test@example.com');
-  });
-
-  it('verifies sortedCountryOptions renders options in alphabetical order', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-    const countrySelect = screen.getByLabelText('country');
-    const options = within(countrySelect).getAllByRole('option');
-
-    // First option is the disabled placeholder
-    expect(options[0].textContent).toContain('select country');
-
-    // Extract country labels from rendered options (skipping the first disabled one)
-    const renderedLabels = options.slice(1).map((opt) => opt.textContent);
-
-    // Sort the original countryOptions by label to compare
-    const sortedOriginalLabels = [...countryOptions]
-      .sort((a, b) => a.label.localeCompare(b.label))
-      .map((opt) => opt.label);
-
-    expect(renderedLabels).toEqual(sortedOriginalLabels);
-  });
-
-  it('exercises onChange for all text inputs', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-
-    const inputs = [
-      { label: 'mobilePhoneNumber', field: 'mobilePhoneNumber', val: '111' },
-      { label: 'workPhoneNumber', field: 'workPhoneNumber', val: '222' },
-      { label: 'homePhoneNumber', field: 'homePhoneNumber', val: '333' },
-      { label: 'addressLine1', field: 'addressLine1', val: 'New Addr 1' },
-      { label: 'addressLine2', field: 'addressLine2', val: 'New Addr 2' },
-      { label: 'postalCode', field: 'postalCode', val: '54321' },
-      { label: 'city', field: 'city', val: 'New City' },
-      { label: 'state', field: 'state', val: 'New State' },
-    ];
-
-    inputs.forEach(({ label, field, val }) => {
-      const input = screen.getByLabelText(label);
-      fireEvent.change(input, { target: { value: val } });
-      expect(mockHandleFieldChange).toHaveBeenCalledWith(field, val);
-    });
-  });
-
-  it('exercises onChange for country select', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-    const countrySelect = screen.getByLabelText('country');
-
-    // Pick a country code from options to simulate change
-    const targetCountry = countryOptions[1].value; // Pick second just to be sure
-    fireEvent.change(countrySelect, { target: { value: targetCountry } });
-
-    expect(mockHandleFieldChange).toHaveBeenCalledWith(
-      'countryCode',
-      targetCountry,
-    );
-  });
-
-  it('checks accessibility attributes', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-
-    // Check htmlFor associations by ensuring querying by label works (which we did above),
-    // but explicit check:
-    expect(screen.getByLabelText('email')).toHaveAttribute('id', 'email');
-
-    // Check aria-label on country options
-    const countrySelect = screen.getByLabelText('country');
-    const options = within(countrySelect).getAllByRole('option');
-    // The component uses tCommon('selectAsYourCountry', { country: label }) for aria-label
-    // Since we mock t -> key, it might come out as 'selectAsYourCountry' or similar depending on how interpolation is handled in the real app vs mock.
-    // In our mock, t returns key. If interpolation is passed, it's ignored by `t: (key) => key`.
-    // Wait, the component passes a second arg for interpolation. The mock should ideally handle it or we just check it returns the key.
-    // The component code: aria-label={tCommon('selectAsYourCountry', { country: country.label })}
-    // Our mock returns 'selectAsYourCountry'.
-
-    expect(options[1]).toHaveAttribute('aria-label', 'selectAsYourCountry');
-  });
-
-  it('supports keyboard navigation (tabbing)', () => {
-    render(<ProfileContactInfo {...defaultProps} />);
-    const emailInput = screen.getByLabelText('email');
-    const mobileInput = screen.getByLabelText('mobilePhoneNumber');
-
-    emailInput.focus();
-    // Since we can't fully simulate tab browser behavior in JSDOM easily without userEvent.tab(),
-    // we assume the DOM order is correct.
-    // We can check valid tabIndices or just reliance on natural DOM order.
-    // The component doesn't set explicit tabIndex except for maybe buttons elsewhere.
-    // Inputs are standard.
-    // We can just assert they are in the document and not hidden.
-    expect(emailInput).toBeVisible();
-    expect(mobileInput).toBeVisible();
-  });
-});
