+++ src/shared-components/ProfileForm/ProfileForm.tsx
@@ -89,19 +89,28 @@ import { USER_DETAILS } from 'GraphQl/Queries/Queries';
 import { NotificationToast } from 'components/NotificationToast/NotificationToast';
 import { languages } from 'utils/languages';
 import { errorHandler } from 'utils/errorHandler';
-import { Card, Row, Col } from 'react-bootstrap';
+import { Card, Row, Col, Form } from 'react-bootstrap';
 import LoadingState from 'shared-components/LoadingState/LoadingState';
 import useLocalStorage from 'utils/useLocalstorage';
+import Avatar from 'components/Avatar/Avatar';
 import EventsAttendedByMember from 'components/MemberActivity/EventsAttendedByMember';
 import MemberAttendedEventsModal from 'components/MemberActivity/Modal/EventsAttendedMemberModal';
+import {
+  countryOptions,
+  educationGradeEnum,
+  maritalStatusEnum,
+  genderEnum,
+  employmentStatusEnum,
+} from 'utils/formEnumFields';
+import dayjs from 'dayjs';
+import DynamicDropDown from 'components/DynamicDropDown/DynamicDropDown';
 import { urlToFile } from 'utils/urlToFile';
 import { validatePassword } from 'utils/passwordValidator';
+import { sanitizeAvatars } from 'utils/sanitizeAvatar';
 import type { IEvent } from 'types/Event/interface';
 import ProfileFormWrapper from './ProfileFormWrapper';
+import DatePicker from 'shared-components/DatePicker';
 import { removeEmptyFields, validateImageFile } from 'utils/userUpdateUtils';
-import ProfilePersonalDetails from './ProfilePersonalDetails';
-import ProfileContactInfo from './ProfileContactInfo';
-import { IProfileFormState } from './types';
 
 const MemberDetail: React.FC = (): JSX.Element => {
   const { t } = useTranslation('translation', { keyPrefix: 'memberDetail' });
@@ -129,7 +138,7 @@ const MemberDetail: React.FC = (): JSX.Element => {
     document.title = t('title');
   }, [t]);
 
-  const [formState, setFormState] = useState<IProfileFormState>({
+  const [formState, setFormState] = useState({
     addressLine1: '',
     addressLine2: '',
     birthDate: null,
@@ -173,14 +182,7 @@ const MemberDetail: React.FC = (): JSX.Element => {
 
   // Function to handle the click on the edit icon
   const handleFileUpload = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    // Use currentTarget for type safety - guaranteed to be HTMLInputElement
-    const { files } = e.currentTarget;
-
-    if (!files || files.length === 0) {
-      return;
-    }
-
-    const file = files[0];
+    const file = e.target?.files?.[0];
 
     const isValid = validateImageFile(file, tCommon);
     if (!isValid || !file) {
@@ -193,14 +195,11 @@ const MemberDetail: React.FC = (): JSX.Element => {
   };
 
   // to handle the change in the form fields
-  const handleFieldChange = (
-    fieldName: keyof IProfileFormState,
-    value: string | File | null,
-  ): void => {
+  const handleFieldChange = (fieldName: string, value: string): void => {
     // future birthdates are not possible to select.
 
     // password validation
-    if (fieldName === 'password' && typeof value === 'string' && value) {
+    if (fieldName === 'password' && value) {
       if (!validatePassword(value)) {
         NotificationToast.error(
           tCommon('passwordLengthRequirement', { length: 8 }) as string,
@@ -310,15 +309,215 @@ const MemberDetail: React.FC = (): JSX.Element => {
       >
         <Row className="g-4 mt-1">
           <Col md={6}>
-            <ProfilePersonalDetails
-              formState={formState}
-              setFormState={setFormState}
-              handleFieldChange={handleFieldChange}
-              selectedAvatar={selectedAvatar}
-              fileInputRef={fileInputRef}
-              handleFileUpload={handleFileUpload}
-              userRole={userData?.user?.role}
-            />
+            {/* Personal Details Card */}
+            <Card className={`${styles.allRound}`}>
+              <Card.Header
+                className={`py-3 px-4 d-flex justify-content-between align-items-center ${styles.topRadius} ${memberDetailStyles.personalDetailsHeader}`}
+              >
+                <h3 className="m-0">{t('personalDetailsHeading')}</h3>
+                <Button
+                  variant="light"
+                  size="sm"
+                  disabled
+                  className="rounded-pill fw-bolder"
+                >
+                  {userData?.user?.role === 'administrator'
+                    ? tCommon('Admin')
+                    : tCommon('User')}
+                </Button>
+              </Card.Header>
+              <Card.Body className="py-3 px-3">
+                <Col lg={12} className="mb-2">
+                  <div className="text-center mb-3">
+                    <div className="position-relative d-inline-block">
+                      {formState?.avatarURL ? (
+                        <img
+                          className={`rounded-circle ${memberDetailStyles.profileImage}`}
+                          src={sanitizeAvatars(
+                            selectedAvatar,
+                            formState.avatarURL,
+                          )}
+                          alt={tCommon('user')}
+                          data-testid="profile-picture"
+                          crossOrigin="anonymous" // to avoid Cors
+                        />
+                      ) : (
+                        <Avatar
+                          name={formState.name}
+                          alt={tCommon('displayImage')}
+                          size={60}
+                          dataTestId="profile-picture"
+                          radius={150}
+                        />
+                      )}
+                      <button
+                        type="button"
+                        className={`fas fa-edit position-absolute border-0 bottom-0 right-0 p-2 bg-white rounded-circle ${memberDetailStyles.editProfileIcon}`}
+                        onClick={() => fileInputRef.current?.click()}
+                        data-testid="uploadImageBtn"
+                        title={`${tCommon('edit')} ${tCommon('profilePicture')}`}
+                        aria-label={`${tCommon('edit')} ${tCommon('profilePicture')}`}
+                        tabIndex={0}
+                        onKeyDown={(e) =>
+                          e.key === 'Enter' && fileInputRef.current?.click()
+                        }
+                      />
+                    </div>
+                  </div>
+                  <Form.Control
+                    accept="image/*"
+                    id="postphoto"
+                    name="photo"
+                    type="file"
+                    className={`${styles.cardControl} ${memberDetailStyles.hiddenFileInput}`}
+                    data-testid="fileInput"
+                    multiple={false}
+                    ref={fileInputRef}
+                    onChange={handleFileUpload}
+                  />
+                </Col>
+                <Row className="g-3">
+                  <Col md={6}>
+                    <label htmlFor="name" className="form-label">
+                      {tCommon('name')}
+                    </label>
+                    <input
+                      id="name"
+                      value={formState.name}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="name"
+                      data-testid="inputName"
+                      onChange={(e) =>
+                        handleFieldChange('name', e.target.value)
+                      }
+                      required
+                      placeholder={tCommon('name')}
+                    />
+                  </Col>
+                  <Col md={6} data-testid="gender">
+                    <label htmlFor="gender" className="form-label">
+                      {t('gender')}
+                    </label>
+                    <DynamicDropDown
+                      formState={formState}
+                      setFormState={setFormState}
+                      fieldOptions={genderEnum}
+                      fieldName="natalSex"
+                      handleChange={(e) =>
+                        handleFieldChange('natalSex', e.target.value)
+                      }
+                    />
+                  </Col>
+                  <Col md={6}>
+                    <label htmlFor="birthDate" className="form-label">
+                      {t('birthDate')}
+                    </label>
+                    <DatePicker
+                      className="w-100"
+                      value={
+                        formState.birthDate ? dayjs(formState.birthDate) : null
+                      }
+                      maxDate={dayjs()}
+                      onChange={(date) => {
+                        if (!date || !dayjs(date).isValid()) {
+                          handleFieldChange('birthDate', '');
+                          return;
+                        }
+                        const picked = dayjs(date);
+                        handleFieldChange(
+                          'birthDate',
+                          picked.format('YYYY-MM-DD'),
+                        );
+                      }}
+                      data-testid="birthDate"
+                      slotProps={{
+                        textField: {
+                          'aria-label': t('birthDate'),
+                        },
+                      }}
+                    />
+                  </Col>
+                  <Col md={6}>
+                    <label htmlFor="grade" className="form-label">
+                      {t('educationGrade')}
+                    </label>
+                    <DynamicDropDown
+                      formState={formState}
+                      setFormState={setFormState}
+                      fieldOptions={educationGradeEnum}
+                      fieldName="educationGrade"
+                      handleChange={(e) =>
+                        handleFieldChange('educationGrade', e.target.value)
+                      }
+                    />
+                  </Col>
+                  <Col md={6}>
+                    <label htmlFor="empStatus" className="form-label">
+                      {t('employmentStatus')}
+                    </label>
+                    <DynamicDropDown
+                      formState={formState}
+                      setFormState={setFormState}
+                      fieldOptions={employmentStatusEnum}
+                      fieldName="employmentStatus"
+                      handleChange={(e) =>
+                        handleFieldChange('employmentStatus', e.target.value)
+                      }
+                    />
+                  </Col>
+                  <Col md={6}>
+                    <label htmlFor="maritalStatus" className="form-label">
+                      {t('maritalStatus')}
+                    </label>
+                    <DynamicDropDown
+                      formState={formState}
+                      setFormState={setFormState}
+                      fieldOptions={maritalStatusEnum}
+                      fieldName="maritalStatus"
+                      handleChange={(e) =>
+                        handleFieldChange('maritalStatus', e.target.value)
+                      }
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="password" className="form-label">
+                      {tCommon('password')}
+                    </label>
+                    <input
+                      id="password"
+                      value={formState.password}
+                      className={`form-control ${styles.inputColor}`}
+                      type="password"
+                      name="password"
+                      onChange={(e) =>
+                        handleFieldChange('password', e.target.value)
+                      }
+                      data-testid="inputPassword"
+                      placeholder="* * * * * * * *"
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="description" className="form-label">
+                      {tCommon('description')}
+                    </label>
+                    <input
+                      id="description"
+                      value={formState.description}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="description"
+                      data-testid="inputDescription"
+                      onChange={(e) =>
+                        handleFieldChange('description', e.target.value)
+                      }
+                      required
+                      placeholder={tCommon('enterDescription')}
+                    />
+                  </Col>
+                </Row>
+              </Card.Body>
+            </Card>
             <Col>
               <Card className={`${styles.contact} ${styles.allRound} mt-3`}>
                 <Card.Header
@@ -378,11 +577,213 @@ const MemberDetail: React.FC = (): JSX.Element => {
           </Col>
 
           <Col md={6}>
-            <ProfileContactInfo
-              formState={formState}
-              userEmail={userData?.user?.emailAddress}
-              handleFieldChange={handleFieldChange}
-            />
+            <Card className={`${styles.allRound}`}>
+              <Card.Header
+                className={`py-3 px-4 ${styles.topRadius} ${memberDetailStyles.member_details_style}`}
+              >
+                <h3 className="m-0">{t('contactInfoHeading')}</h3>
+              </Card.Header>
+              <Card.Body className="py-3 px-3">
+                <Row className="g-3">
+                  <Col md={12}>
+                    <label htmlFor="email" className="form-label">
+                      {tCommon('email')}
+                    </label>
+                    <input
+                      id="email"
+                      value={userData?.user?.emailAddress}
+                      className={`form-control ${styles.inputColor}`}
+                      type="email"
+                      name="email"
+                      data-testid="inputEmail"
+                      disabled
+                      placeholder={tCommon('email')}
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="mobilePhoneNumber" className="form-label">
+                      {t('mobilePhoneNumber')}
+                    </label>
+                    <input
+                      id="mobilePhoneNumber"
+                      value={formState.mobilePhoneNumber}
+                      className={`form-control ${styles.inputColor}`}
+                      type="tel"
+                      name="mobilePhoneNumber"
+                      data-testid="inputMobilePhoneNumber"
+                      onChange={(e) =>
+                        handleFieldChange('mobilePhoneNumber', e.target.value)
+                      }
+                      placeholder={
+                        tCommon('example', { example: '+1234567890' }) as string
+                      }
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="workPhoneNumber" className="form-label">
+                      {t('workPhoneNumber')}
+                    </label>
+                    <input
+                      id="workPhoneNumber"
+                      value={formState.workPhoneNumber}
+                      className={`form-control ${styles.inputColor}`}
+                      type="tel"
+                      data-testid="inputWorkPhoneNumber"
+                      name="workPhoneNumber"
+                      onChange={(e) =>
+                        handleFieldChange('workPhoneNumber', e.target.value)
+                      }
+                      placeholder={
+                        tCommon('example', { example: '+1234567890' }) as string
+                      }
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="homePhoneNumber" className="form-label">
+                      {t('homePhoneNumber')}
+                    </label>
+                    <input
+                      id="homePhoneNumber"
+                      value={formState.homePhoneNumber}
+                      className={`form-control ${styles.inputColor}`}
+                      type="tel"
+                      data-testid="inputHomePhoneNumber"
+                      name="homePhoneNumber"
+                      onChange={(e) =>
+                        handleFieldChange('homePhoneNumber', e.target.value)
+                      }
+                      placeholder={
+                        tCommon('example', { example: '+1234567890' }) as string
+                      }
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="addressLine1" className="form-label">
+                      {t('addressLine1')}
+                    </label>
+                    <input
+                      id="addressLine1"
+                      value={formState.addressLine1}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="addressLine1"
+                      data-testid="addressLine1"
+                      onChange={(e) =>
+                        handleFieldChange('addressLine1', e.target.value)
+                      }
+                      placeholder={
+                        tCommon('example', { example: 'Lane 1' }) as string
+                      }
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="addressLine2" className="form-label">
+                      {t('addressLine2')}
+                    </label>
+                    <input
+                      id="addressLine2"
+                      value={formState.addressLine2}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="addressLine2"
+                      data-testid="addressLine2"
+                      onChange={(e) =>
+                        handleFieldChange('addressLine2', e.target.value)
+                      }
+                      placeholder={
+                        tCommon('example', { example: 'Lane 2' }) as string
+                      }
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <label htmlFor="postalCode" className="form-label">
+                      {t('postalCode')}
+                    </label>
+                    <input
+                      id="postalCode"
+                      value={formState.postalCode}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="postalCode"
+                      data-testid="inputPostalCode"
+                      onChange={(e) =>
+                        handleFieldChange('postalCode', e.target.value)
+                      }
+                      placeholder={
+                        tCommon('example', {
+                          example: '12345',
+                        }) as string
+                      }
+                    />
+                  </Col>
+                  <Col md={6}>
+                    <label htmlFor="city" className="form-label">
+                      {t('city')}
+                    </label>
+                    <input
+                      id="city"
+                      value={formState.city}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="city"
+                      data-testid="inputCity"
+                      onChange={(e) =>
+                        handleFieldChange('city', e.target.value)
+                      }
+                      placeholder={tCommon('enterCityName')}
+                    />
+                  </Col>
+                  <Col md={6}>
+                    <label htmlFor="state" className="form-label">
+                      {t('state')}
+                    </label>
+                    <input
+                      id="state"
+                      value={formState.state}
+                      className={`form-control ${styles.inputColor}`}
+                      type="text"
+                      name="state"
+                      data-testid="inputState"
+                      onChange={(e) =>
+                        handleFieldChange('state', e.target.value)
+                      }
+                      placeholder={tCommon('enterStateName')}
+                    />
+                  </Col>
+                  <Col md={12}>
+                    <Form.Label htmlFor="country" className="form-label">
+                      {tCommon('country')}
+                    </Form.Label>
+                    <Form.Select
+                      id="country"
+                      value={formState.countryCode}
+                      className={`${styles.inputColor}`}
+                      data-testid="inputCountry"
+                      onChange={(e) =>
+                        handleFieldChange('countryCode', e.target.value)
+                      }
+                    >
+                      <option value="" disabled>
+                        {tCommon('select')} {tCommon('country')}
+                      </option>
+                      {[...countryOptions]
+                        .sort((a, b) => a.label.localeCompare(b.label))
+                        .map((country) => (
+                          <option
+                            key={country.value.toUpperCase()}
+                            value={country.value.toLowerCase()}
+                            aria-label={tCommon('selectAsYourCountry', {
+                              country: country.label,
+                            })}
+                          >
+                            {country.label}
+                          </option>
+                        ))}
+                    </Form.Select>
+                  </Col>
+                </Row>
+              </Card.Body>
+            </Card>
           </Col>
 
           {isUpdated && (
